Phase 4: party state machine, lifecycle, and protocol coverage
Bundles weeks of incremental work into a single integration commit since
nothing in here was independently shippable. Highlights:

* Party state machine — DpsAggregator's PartyRosterUpdate handler now
  decides every roster decision via (rosterRoom vs _currentMatchmakingRoom,
  selfInRoster) → UPDATE / KICKED / ROOM_CHANGE / BROWSE_IGNORE. Replaces
  the prior pile of host-flicker / weak-removal / ADD-only guards that
  kept conflicting with each other. Cross-checked against A2Viewer
  PartyTracker source.
* Cold-start bootstrap — first Strong (op=02 97) before SelfUserId is
  inferred is trusted as the user's room (Strong only fires for the
  user's own room per A2Viewer RE).
* Eviction scope — EvictStaleMembers only runs when the user is not in
  any matchmaking room. Inside a room (lobby or dungeon-from-room) the
  state machine owns membership; eviction was wiping non-self members
  60s after dungeon entry due to sparse OTHER_NICK enrichment.
* Canonical-id Player model — NicknameRegistry aliases all id-spaces
  (lobby / dungeon / SELF_NICK / raid bulk) to one canonical entity_id
  per nickname. Eliminates the duplicate-row / partial-CP class of bugs.
* Concurrency — DpsAggregator wraps OnEvent / OurCrew / EvictStaleMembers
  in a state lock; Session._players moved to ConcurrentDictionary. UI
  thread no longer races the capture thread.
* Channel capacity — bumped 16384 → 65536 after wire-confirmed 21k
  packet drops in 30s during lobby browse + room creation, which had
  silently dropped joiner notifications.
* Boss / dungeon coverage — BossTracker HP fields long, DungeonDatabase
  + ServerMap, EncounterAnnouncement → mob_code linking, post-kill
  display hold via LastKilledBossId.
* Protocol handlers — full op=__ 97 family (PartyAssembly 01/02, Member
  Status 0B / friend pings, PartyLeft 1D), op=00 92 CP refresh,
  op=01 91 EncounterAnnounce, op=40 36 SummonSpawn.
* UI — class icons + Korean class names, [me] tag, primary-actor
  heuristic with placeholder row, manual drag (no Aero Snap), foreground
  watcher for game-overlay show/hide, ResetSession command.

+++ Aion2FunDps.Core/Models/MobHpUpdate.cs
@@ -2,7 +2,7 @@ namespace Aion2FunDps.Core.Models;
 
 public sealed record MobHpUpdate(
     int MobId,
-    uint CurrentHp,
+    long CurrentHp,
     long TimestampTicks,
     uint SourceIpv4
 ) : IGameEvent;
